add support for let-expression
add support for let-expression. let here acts the same as let* in
racket. i.e. the location is created as soon as the val-expr is
successfully evaluated and ids need not to be distinct. Later bindings
shallow old ones.

fix bug, (define f (lmabda)) would cause seg-fault in old implemetation

resolve the issue that the function cannot be recusively called if its
name is the same as that of one of its parameters. Those functions will
simply not allowed to be defined.

change the design decision for closure. I adopt the same method as the
used in SMLNJ interpreter.
i.e. (define a 2)
     (define (f x) (+ a x))
     (f 3) --> 5
     (define a 4)
     (f 3) --> 5
therefore, every variable used in closure should be defined forward
the following code (allowed in python)  should *not* be allowed here.
     (define (f x) (+ a x))
     (define a 2)
     (f 2) -- > 4